refactor: WIP on using the new constants
