+++ yarn-project/circuits.js/src/cbind/constants.gen.ts
@@ -46,6 +46,7 @@ export const NULLIFIER_SUBTREE_SIBLING_PATH_LENGTH = NULLIFIER_TREE_HEIGHT - NUL
 export const L1_TO_L2_MSG_SUBTREE_HEIGHT = log2(NUMBER_OF_L1_L2_MESSAGES_PER_ROLLUP);
 export const L1_TO_L2_MSG_SUBTREE_SIBLING_PATH_LENGTH = L1_TO_L2_MSG_TREE_HEIGHT - L1_TO_L2_MSG_SUBTREE_HEIGHT;
 export const FUNCTION_SELECTOR_NUM_BYTES = 4;
+export const MAPPING_SLOT_PEDERSEN_SEPARATOR = 4;
 export const NUM_FIELDS_PER_SHA256 = 2;
 export enum GeneratorIndex {
   COMMITMENT = 1,
