+++ yarn-project/circuits.js/src/structs/index.ts
@@ -6,7 +6,6 @@ export * from './rollup/append_only_tree_snapshot.js';
 export * from './rollup/base_or_merge_rollup_public_inputs.js';
 export * from './rollup/previous_rollup_data.js';
 export * from './call_context.js';
-export * from './constants.js';
 export * from './function_data.js';
 export * from './kernel/private_kernel.js';
 export * from './kernel/public_kernel.js';
